requirements/development.txt
-r ./base.txt

flake8
pytest
pytest-django
django-test-plus